backend/requirements.txt
fastapi[all]
sqlalchemy
email-validator
requests
apscheduler
alembic
python-dotenv
passlib[bcrypt]
langchain-gigachat